publishing(#9) : 메인 라우터 수정
