+++ front/src/App.js
@@ -3,7 +3,6 @@ import HomePage from './components/HomePage/HomePage.js';
 import Login from './components/HomePage/Login.js';
 import SignIn from './components/HomePage/SignIn.js';
 import Main from './components/Main/Main.js';
-import NewPost from './components/Main/NewPost.js';
 
 function App() {
   return (
@@ -13,8 +12,9 @@ function App() {
           <Route exact path="/" Component={HomePage} />
           <Route exact path="/login" Component={Login} />
           <Route exact path="/sign-in" Component={SignIn} />
-          <Route exact path="/norang" Component={Main} />
-          <Route exact path="/new-post" Component={NewPost} />
+          <Route path="/NORANG" element={<Main page="postList" />} />
+          <Route path="/new-post" element={<Main page="newPost" />} />
+          <Route path="/detail-post" element={<Main page="detailPost" />} />
         </Routes>
       </BrowserRouter>
     </div>
