+++ front/src/pages/styles/index.css
@@ -18,6 +18,7 @@
   margin: 0;
   font-family: 'Cafe24Ssurround';
   color: #606060;
+  text-decoration-line: none;
 }
 
 .body {
@@ -254,4 +255,8 @@ label {
   font-family: 'Cafe24SsurroundAir';
   padding: 0vw 2.5vw 2.5vw 2.5vw;
   font-size: 85%;
+}
+
+.post:hover {
+  font-weight: bolder;
 }
